+++ bindings/dotnet/src/DecentDB.EntityFrameworkCore/Query/Internal/DecentDBRelationalParameterBasedSqlProcessor.cs
@@ -1,0 +1,19 @@
+using System.Linq.Expressions;
+using Microsoft.EntityFrameworkCore.Query;
+
+namespace DecentDB.EntityFrameworkCore.Query.Internal;
+
+public sealed class DecentDBRelationalParameterBasedSqlProcessor : RelationalParameterBasedSqlProcessor
+{
+    public DecentDBRelationalParameterBasedSqlProcessor(
+        RelationalParameterBasedSqlProcessorDependencies dependencies,
+        RelationalParameterBasedSqlProcessorParameters parameters)
+        : base(dependencies, parameters)
+    {
+    }
+
+    protected override Expression ProcessSqlNullability(
+        Expression queryExpression,
+        ParametersCacheDecorator parametersDecorator)
+        => new DecentDBSqlNullabilityProcessor(Dependencies, Parameters).Process(queryExpression, parametersDecorator);
+}